requirements.txt
# =============================================================
# RoentGen – Bağımlılıklar (Makaledeki ortamla uyumlu)
# =============================================================
# Kurulum:  pip install -r requirements.txt
# Not: PyTorch'u CUDA sürümünüze göre yüklemek için:
#   pip install torch==1.12.1+cu116 torchvision==0.13.1+cu116 \
#       --extra-index-url https://download.pytorch.org/whl/cu116
# =============================================================

# --- Makalede belirtilen sabit sürümler ---
diffusers==0.7.1
vilmedic==1.2.8
numpy==1.23.1
scipy==1.9.1
pandas==1.4.4
torchmetrics==0.11.0

# --- PyTorch (diffusers 0.7.1 ile uyumlu kararlı sürümler) ---
torch==1.12.1
torchvision==0.13.1

# --- Hugging Face ekosistemi ---
transformers==4.25.1
accelerate==0.15.0
tokenizers==0.13.2

# --- Görüntü işleme & yardımcı ---
Pillow==9.2.0
ftfy==6.1.1
